pyproject.toml
[tool.poetry]
name = "chatgpt-example"
version = "0.1.0"
description = ""
authors = ["daylight55 <jojomasanori3@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
